refactor: sidebar button 로직 hook으로 분리 & 버튼 상태 disabled 속성으로 반영
