+++ frontend/src/business/hooks/useSideBar/index.ts
@@ -1,0 +1,1 @@
+export { useSideBarButton } from './useSideBarButton';
